move schema stuff out of write util functions
