+++ server/routes/songs/util/write.js
@@ -1,92 +1,10 @@
 const database = require('../../../db');
 
-// TODO `required` flags in parity w/ sql
-const songsWriteSchema = {
-  fields: {
-    name: { required: true },
-    description: { required: true },
-    imageUrl: { db: 'image_url', required: true },
-    songFileName: { db: 'song_file_name', required: true },
-    curatorId: { db: 'curator_id', required: true },
-    artistName: { db: 'artist_name', required: true },
-    spotify: {
-      required: true, // DB way ... shouldn't we actually at least one / exactly one link type?
-      fields: {
-        link: { required: true },
-      }
-    },
-    soundcloud: {
-      fields: {
-        link: { required: true },
-        trackId: { db: 'track_id', required: true },
-      }
-    },
-    youtube: {
-      fields: {
-        link: { required: true },
-        trackId: { db: 'track_id', required: true },
-      }
-    },
-    // Should these really be required?
-    bpm: { required: true },
-    artistLocation: { db: 'artist_location', required: true },
-    createdAt: { db: 'created_at', required: true },
-    hidden: {}
-  },
-  relations: {
-    subgenreIds: { join_table: 'subgenre_songs', db: 'subgenre_id' },
-    momentIds: { join_table: 'song_moments', db: 'moment_id' }
-  }
-}
-
-
-// TODO import recompose and use this w/ `compose` in `getMissingRequiredFields` instead of reduce
-const keyValueArrayToObject = keyValueArray => (
-    keyValueArray.reduce( (curr, [ key, value ]) => ({
-      ...curr,
-      key: value
-    }), {})
-);
-
-
-const requiredFieldNameIfMissing = (fieldName, fieldData, params) => (
-  (!params[fieldName] && fieldData.required) ?
-    fieldName :
-    (!!fieldData.fields && !!params[fieldName] && fieldData.isRequired) ?
-      getMissingRequiredFields(params[fieldName], fieldData.fields) :
-      null
-);
-
-const getMissingRequiredFields = (params, schema = songsWriteSchema.fields) => (
-    Object.entries(schema)
-      .map( ([ fieldName, fieldData ]) => (
-          [ fieldName, requiredFieldNameIfMissing(fieldName, fieldData, params) ]
-      ))
-      .filter( ([fieldName, fieldData ]) => fieldData != null && Object.keys(fieldData).length > 0 )
-      .reduce( (allMissingFields, [ fieldName, missingData ]) => ({
-        ...allMissingFields,
-        [fieldName]: missingData
-      }), {})
-);
-
-const getDbFieldValues = (params, schema = songsWriteSchema.fields, dbNamePrefix = '') => (
-    Object.entries(schema)
-      .reduce ( (currPairs, [fieldName, fieldData]) => {
-
-        const paramValue = (!params[fieldName] && !!fieldData.default) ?
-          fieldData.default() :
-          params[fieldName];
-
-        const dbFieldName = dbNamePrefix + (fieldData.db ? fieldData.db : fieldName);
-
-        const nextPairs = (!!fieldData.fields && !!paramValue) ?
-          getDbFieldValues(params[fieldName], fieldData.fields, fieldName + '_') :
-          [ [dbFieldName, paramValue] ];
-
-        return [ ...currPairs, ...nextPairs ];
-      }, [] )
-      .filter( ([fieldName, fieldData ]) => !!fieldData )
-);
+const {
+  getMissingRequiredFields,
+  getDbFieldValues,
+  getAllRelationQueries
+} = require('./schema.js');
 
 const insertSong = params => {
   const missingFields = getMissingRequiredFields(params);
@@ -111,31 +29,6 @@ const insertSong = params => {
   })
   .then(insertResponse => insertRelation(insertResponse.rows[0].id, params));
 };
-
-const processRelationQueries = getQuery =>
-  Promise.all(
-    Object.keys(songsWriteSchema.relations).map(relationName => {
-      if (!params[relationName] || params[relationName].length == 0)
-        return null;
-
-      const { joinTable, db } = songsWriteSchema.relations[relationName];
-
-      return getQuery(relationName, joinTable, db).then(database.query);
-    })
-  );
-
-const insertRelations = (songId, params) => {
-  if (!songId)
-    throw "Cannot insert relations w/o song id";
-
-  return processRelationQueries((relationName, joinTable, db) => ({
-    text: `
-      INSERT INTO ${joinTable} (song_id, ${db}) VALUES ' +
-      params[relationName].map((id, i) => '($1,$' + (i+2) + ')').join(',');
-    `,
-    values: [ songId, ...params[relationName] ]
-  }));
-}
 
 const updateSong = (songId, params) => {
   if (!songId) {
@@ -174,11 +67,27 @@ const deleteSong = songId => {
   .then(() => deleteRelations(songId));
 }
 
+const performRelationQueries = getQuery =>
+  Promise.all(getAllRelationQueries.map(database.query))
+
+const insertRelations = (songId, params) => {
+  if (!songId)
+    throw "Cannot insert relations w/o song id";
+
+  return performRelationQueries((relationName, joinTable, db) => ({
+    text: `
+      INSERT INTO ${joinTable} (song_id, ${db}) VALUES ' +
+      params[relationName].map((id, i) => '($1,$' + (i+2) + ')').join(',');
+    `,
+    values: [ songId, ...params[relationName] ]
+  }));
+}
+
 const deleteRelations = songId => {
   if (!songId)
     throw "Cannot delete relations w/o song id";
 
-  return processRelationQueries((relationName, joinTable, db) => ({
+  return performRelationQueries((relationName, joinTable, db) => ({
     text: `DELETE FROM ${join_table} WHERE subgenre_songs.song_id = $1`,
     values: [ songId ]
   }));
